# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Creative Room API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  clerkId: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  musicalStyle?: string | null;
  /** @nullable */
  emotionalVibe?: string | null;
  /** @nullable */
  inspirations?: string | null;
  genres?: string[];
  /** @nullable */
  avatarUrl?: string | null;
  isCreator?: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface ProfileInput {
  /** @minLength 1 */
  displayName: string;
  bio?: string;
  musicalStyle?: string;
  emotionalVibe?: string;
  inspirations?: string;
  genres?: string[];
  avatarUrl?: string;
}

export interface ProfileUpdate {
  /** @minLength 1 */
  displayName?: string;
  bio?: string;
  musicalStyle?: string;
  emotionalVibe?: string;
  inspirations?: string;
  genres?: string[];
  avatarUrl?: string;
}

export interface Room {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  vibe?: string | null;
  genres?: string[];
  maxMembers: number;
  ownerId: number;
  /** @nullable */
  ownerName?: string | null;
  isActive: boolean;
  memberCount: number;
  /** @nullable */
  coverImageUrl?: string | null;
  createdAt: string;
}

/**
 * Map of roomId (as string key) to the number of live connections in that room
 */
export interface RoomPresenceMap {[key: string]: number}

export interface RoomInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  vibe?: string;
  genres?: string[];
  /**
     * @minimum 2
     * @maximum 8
     */
  maxMembers: number;
  coverImageUrl?: string;
}

export interface RoomUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  vibe?: string;
  genres?: string[];
  maxMembers?: number;
  isActive?: boolean;
  coverImageUrl?: string;
}

export type RoomMemberRole = typeof RoomMemberRole[keyof typeof RoomMemberRole];


export const RoomMemberRole = {
  owner: 'owner',
  member: 'member',
} as const;

export interface RoomMember {
  id: number;
  roomId: number;
  profileId: number;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: RoomMemberRole;
  joinedAt: string;
}

export interface Message {
  id: number;
  roomId: number;
  profileId: number;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  senderAvatarUrl?: string | null;
  content: string;
  createdAt: string;
}

export interface Demo {
  id: number;
  roomId: number;
  uploaderId: number;
  /** @nullable */
  uploaderName?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  fileUrl: string;
  /** @nullable */
  duration?: number | null;
  createdAt: string;
}

export interface RoomActivity {
  roomId: number;
  messageCount: number;
  demoCount: number;
  memberCount: number;
  /** @nullable */
  lastActivityAt: string | null;
  recentMessages?: Message[];
  recentDemos?: Demo[];
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface DemoInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  fileUrl: string;
  duration?: number;
}

export interface Hook {
  id: number;
  creatorId: number;
  /** @nullable */
  creatorName?: string | null;
  /** @nullable */
  creatorAvatarUrl?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  audioUrl: string;
  /** @nullable */
  vibe?: string | null;
  tags?: string[];
  lookingFor: string[];
  maxSeats: number;
  seatsLeft: number;
  /** @nullable */
  roomId?: number | null;
  isActive: boolean;
  createdAt: string;
}

export interface HookUpdateInput {
  isActive?: boolean;
}

export interface HookInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  audioUrl: string;
  vibe?: string;
  tags?: string[];
  lookingFor?: string[];
  /**
     * @minimum 2
     * @maximum 4
     */
  maxSeats?: number;
}

export type ListRoomsParams = {
genre?: string;
vibe?: string;
};