# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Creative Room API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms/{id}/demos:
    get:
      operationId: getRoomDemos
      tags: [demos]
      summary: Get demos uploaded to a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Demos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Demo"
    post:
      operationId: uploadDemo
      tags: [demos]
      summary: Upload a demo to a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoInput"
      responses:
        "201":
          description: Demo uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demo"
  /rooms/{roomId}/demos/{demoId}:
    delete:
      operationId: deleteDemo
      tags: [demos]
      summary: Delete a demo
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
        - name: demoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Hooks ─────────────────────────────────────────────────────────────────
  /hooks:
    get:
      operationId: listHooks
      tags: [hooks]
      summary: List active hooks
      responses:
        "200":
          description: List of hooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hook"
    post:
      operationId: createHook
      tags: [hooks]
      summary: Drop a new hook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hook"
  /hooks/{id}:
    get:
      operationId: getHook
      tags: [hooks]
      summary: Get a hook by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hook found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hook"
        "404":
          description: Not found
    patch:
      operationId: updateHook
      tags: [hooks]
      summary: Update a hook (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HookUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hook"
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteHook
      tags: [hooks]
      summary: Delete a hook (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
        "404":
          description: Not found
  /hooks/{id}/join:
    post:
      operationId: joinHook
      tags: [hooks]
      summary: Join a hook — creates or joins the associated room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hook"
        "400":
          description: Hook is full or inactive
        "404":
          description: Not found
components:
  schemas:
...
    Demo:
      type: object
      required: [id, roomId, uploaderId, title, fileUrl, createdAt]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        uploaderId:
          type: integer
        uploaderName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        fileUrl:
          type: string
        duration:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    DemoInput:
      type: object
      required: [title, fileUrl]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        fileUrl:
          type: string
        duration:
          type: integer

    # ── Hook ─────────────────────────────────────────────────────────────────
    Hook:
      type: object
      required: [id, creatorId, title, audioUrl, lookingFor, maxSeats, seatsLeft, isActive, createdAt]
      properties:
        id:
          type: integer
        creatorId:
          type: integer
        creatorName:
          type: ["string", "null"]
        creatorAvatarUrl:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        audioUrl:
          type: string
        vibe:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        lookingFor:
          type: array
          items:
            type: string
        maxSeats:
          type: integer
        seatsLeft:
          type: integer
        roomId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    HookUpdateInput:
      type: object
      properties:
        isActive:
          type: boolean
    HookInput:
      type: object
      required: [title, audioUrl]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        audioUrl:
          type: string
        vibe:
          type: string
        tags:
          type: array
          items:
            type: string
        lookingFor:
          type: array
          items:
            type: string
        maxSeats:
          type: integer
          minimum: 2
          maximum: 4